requirements.txt
pandas==1.5.0
protobuf==4.21.6
python-dotenv==0.21.0
PyYAML==6.0
db-dtypes==1.0.4
google-cloud-bigquery==3.3.2
pytest==7.1.3
darts==0.21.0
